requirements.txt
streamlit
mysql-connector-python
python-dotenv
pandas
plotly
streamlit-option-menu
openai
fpdf2